+++ internal/cmd/base.go
@@ -66,10 +66,7 @@ var baseTestCmd = &cobra.Command{Use: "test", RunE: func(cmd *cobra.Command, arg
 func init() {
 	baseAngleCmd.Flags().Int("head", 0, "head angle")
 	baseAngleCmd.Flags().Int("foot", 0, "foot angle")
-	viper.BindPFlag("head", baseAngleCmd.Flags().Lookup("head"))
-	viper.BindPFlag("foot", baseAngleCmd.Flags().Lookup("foot"))
 	basePresetRunCmd.Flags().String("name", "", "preset name")
-	viper.BindPFlag("name", basePresetRunCmd.Flags().Lookup("name"))
 
 	baseCmd.AddCommand(baseInfoCmd, baseAngleCmd, basePresetsCmd, basePresetRunCmd, baseTestCmd)
 }